Introduced IFileNameGenerator
